Update work in progress

+++ Patchr/UIView+BorderLayer.h
@@ -12,6 +12,6 @@
 
 @property (nonatomic) IBInspectable UIColor *borderColor;
 @property (nonatomic) IBInspectable CGFloat borderWidth;
-@property (nonatomic) IBInspectable NSInteger cornerRadius;
+@property (nonatomic) IBInspectable CGFloat cornerRadius;
 
 @end
